+++ sdk/ai/azure-ai-agents/src/samples/java/com/azure/ai/agents/toolboxes/GetToolbox.java
@@ -23,7 +23,7 @@ import com.azure.identity.DefaultAzureCredentialBuilder;
 public class GetToolbox {
     public static void main(String[] args) {
         String endpoint = Configuration.getGlobalConfiguration().get("FOUNDRY_PROJECT_ENDPOINT");
-        String toolboxName = "toolbox-created-from-java";
+        String toolboxName = "toolbox_created_from_java";
         // Code sample for retrieving a toolbox
         ToolboxesClient toolboxesClient = new AgentsClientBuilder()
                 .credential(new DefaultAzureCredentialBuilder().build())