+++ sdk/ai/azure-ai-projects/src/main/java/com/azure/ai/projects/BetaSkillsAsyncClient.java
@@ -705,7 +705,7 @@ public final class BetaSkillsAsyncClient {
     public Mono<SkillVersion> createSkillVersionFromFiles(String name, CreateSkillVersionFromFilesBody content) {
         // Generated convenience method for createSkillVersionFromFilesWithResponseInternal
         RequestOptions requestOptions = new RequestOptions();
-        return createSkillVersionFromFilesWithResponseInternal(name,
+        return createSkillVersionFromFilesInternalWithResponse(name,
             new MultipartFormDataHelper(requestOptions)
                 .serializeFileFields("files",
                     content.getFiles().stream().map(SkillFileDetails::getContent).collect(Collectors.toList()),
@@ -908,7 +908,7 @@ public final class BetaSkillsAsyncClient {
      */
     @Generated
     @ServiceMethod(returns = ReturnType.SINGLE)
-    Mono<Response<BinaryData>> createSkillVersionFromFilesWithResponseInternal(String name, BinaryData content,
+    Mono<Response<BinaryData>> createSkillVersionFromFilesInternalWithResponse(String name, BinaryData content,
         RequestOptions requestOptions) {
         // Operation 'createSkillVersionFromFiles' is of content-type 'multipart/form-data'. Protocol API is not usable
         // and hence not generated.