+++ sdk/ai/azure-ai-projects/src/main/java/com/azure/ai/projects/models/AzureAIAgentTarget.java
@@ -17,7 +17,7 @@ import java.util.List;
  * Represents a target specifying an Azure AI agent.
  */
 @Fluent
-public final class AzureAIAgentTarget extends Target {
+public final class AzureAIAgentTarget extends FoundryEvaluationTarget {
 
     /*
      * The type of target.